pyproject.toml
[tool.poetry]
name = "assistant"
version = "0.1.0"
description = ""
authors = ["Woodenman23 <josephrfoster@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.12.3"
pytz = "^2024.1"
black = "^24.4.2"
requests = "^2.32.3"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
virtual-assistant = "assistant.app:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
